Change VST license to BSD-2-Clause
See also p.r. #2402 which contains this commit comment, which also applies here:

It was already the case that the VST opam distribution was licensed by BSD-2-Clause, as you can see in the VST repo in the file LICENSE-OPAM which has existed for some time now. Other parts of the VST repo, which are not distributed in the opam distribution, may have other licenses, as described in the VST repo in file LICENSE.